+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Function `copy_file_to_dir` in `gcpy/util.py`.  This is a wrapper for `shutil.copyfile`.
 - Script `gcpy/benchmark/modules/benchmark_mass_cons_table.py`, with code to create mass conservation tables
 - Expanded statistics output in benchmark mass conservation tables
+- Function `get_datetimes_from_filenames` in `gcpy/benchmark/modules/benchmark_utils.py`
 
 ### Changed
 - Bump pip from 23.2.1 to 23.3 (dependabot suggested this)